Code refactoring
Logic for categorising and storing emails has been moved out of the main
outlook addin class into its own class. Further refactoring could be
done (e.g. a class which contains the logic for handling emails and
another for saving them to the database). A small class for converting a
MailItem into an EmailClass is also added.